+++ src/RaptorQ/v1/Precode_Matrix_Solver.hpp
@@ -185,7 +185,8 @@ void Precode_Matrix<IS_OFFLINE>::decode_phase0 (const Bitmask &mask,
 	// but we still need to do the same modifications to A also for repair
 	// symbols. And those have been compacted.
 
-	for (uint16_t rep_row = static_cast<uint16_t> (A.rows() - _repair_overhead);
+	for (uint16_t rep_row = static_cast<uint16_t> (
+							static_cast<uint32_t>(A.rows()) - _repair_overhead);
 												rep_row < A.rows(); ++rep_row) {
 		auto depends = _params.get_idxs (static_cast<uint16_t> (
 															*r_esi + padding));
