+++ src/RaptorQ/v1/caches.cpp
@@ -62,24 +62,24 @@ bool set_compression (const Compress compression)
 	return false;
 }
 
-uint64_t shared_cache_size (const uint64_t shared_cache)
+size_t shared_cache_size (const size_t shared_cache)
 {
     return 0;
 }
 
-bool local_cache_size (const uint64_t local_cache)
+bool local_cache_size (const size_t local_cache)
 {
     return RaptorQ__v1::Impl::DLF<std::vector<uint8_t>,
 									RaptorQ__v1::Impl::Cache_Key>::
 												get()->resize (local_cache);
 }
 
-uint64_t get_shared_cache_size()
+size_t get_shared_cache_size()
 {
     return 0;
 }
 
-uint64_t get_local_cache_size()
+size_t get_local_cache_size()
 {
     return RaptorQ__v1::Impl::DLF<std::vector<uint8_t>,
 									RaptorQ__v1::Impl::Cache_Key>::
@@ -130,16 +130,16 @@ Compress get_compression()
 bool set_compression (const Compress compression)
 { return RaptorQ__v1::set_compression (compression); }
 
-uint64_t shared_cache_size (const uint64_t shared_cache)
+size_t shared_cache_size (const size_t shared_cache)
 { return RaptorQ__v1::shared_cache_size (shared_cache); }
 
-bool local_cache_size (const uint64_t local_cache)
+bool local_cache_size (const size_t local_cache)
 { return RaptorQ__v1::local_cache_size (local_cache); }
 
-uint64_t get_shared_cache_size()
+size_t get_shared_cache_size()
 { return RaptorQ__v1::get_shared_cache_size(); }
 
-uint64_t get_local_cache_size()
+size_t get_local_cache_size()
 { return RaptorQ__v1::get_local_cache_size(); }
 
 } // namespace RFC6330__v1
